♻️ Refactor : 코드 분할 & 디렉토리 구조 변경
