+++ src/settings.ts
@@ -1,0 +1,16 @@
+export const DEFAULT_SETTINGS: EasyTestSettings = {
+	koColor: "sample",
+	koBackgroundColor: "sample",
+	enColor: "sample",
+	enBackgroundColor: "sample",
+	numColor: "sample",
+	numBackgroundColor: "sample",
+};
+export interface EasyTestSettings {
+	koColor: string;
+	koBackgroundColor: string;
+	enColor: string;
+	enBackgroundColor: string;
+	numColor: string;
+	numBackgroundColor: string;
+}
